+++ src/proton/shapes/circle.hh
@@ -27,7 +27,7 @@ namespace Proton
       this->y = y;
     }
 
-    void paint(SDL_Renderer *render, const int rX, const int rY) override
+    void paint(SDL_Renderer *render, const float rX, const float rY) override
     {
       switch (this->style) {
         case Bevel: {
@@ -35,8 +35,8 @@ namespace Proton
           for (int angle = 0; angle < 360; ++angle)
           {
             const float rad = angle * M_PI / 180.0f;
-            const int dx = static_cast<int>(radius * std::cos(rad));
-            const int dy = static_cast<int>(radius * std::sin(rad));
+            const float dx = radius * std::cos(rad);
+            const float dy = radius * std::sin(rad);
             SDL_RenderPoint(render, this->x + rX + dx, this->y + rY + dy);
           }
           break;
@@ -46,7 +46,7 @@ namespace Proton
           SDL_SetRenderDrawColor(render, color.getR(), color.getG(), color.getB(), color.getA());
           for (int dy = -radius; dy <= radius; dy++)
           {
-            const int dx = static_cast<int>(std::sqrt(radius * radius - dy * dy));
+            const float dx = std::sqrt(radius * radius - dy * dy);
             SDL_RenderLine(render, (this->x + rX) - dx, (this->y + rY) + dy, (this->x + rX) + dx, (this->y + rY) + dy);
           }
           break;
