+++ src/proton/shapes/line.hh
@@ -26,7 +26,7 @@ namespace Proton
       this->endY = dy + y;
     }
 
-    void paint(SDL_Renderer *render, const int rX, const int rY) override
+    void paint(SDL_Renderer *render, const float rX, const float rY) override
     {
       SDL_SetRenderDrawColor(render, fillColor.getR(), fillColor.getG(),
                              fillColor.getB(), fillColor.getA());
