+++ src/proton/shapes/rectanglebutton.hh
@@ -37,7 +37,7 @@ namespace Proton
             rectangle->resize(w, h);
         }
 
-        void paint(SDL_Renderer *render, const int rX, const int rY) override
+        void paint(SDL_Renderer *render, const float rX, const float rY) override
         {
             rectangle->paint(render, rX, rY);
         }
