+++ src/proton/shapes/textbox.hh
@@ -247,7 +247,7 @@ namespace Proton
             }
         }
 
-        void paint(SDL_Renderer *render, const int rX, const int rY) override
+        void paint(SDL_Renderer *render, const float rX, const float rY) override
         {
             if (this->isDirty)
                 this->createTexture(render);
